Split parser methods into `ParserExtensions"
