+++ modules/recheck/shared/src/main/scala/codes/quine/labo/recheck/fuzz/FuzzProgram.scala
@@ -4,6 +4,7 @@ import scala.util.Try
 
 import codes.quine.labo.recheck.common.Context
 import codes.quine.labo.recheck.regexp.Pattern
+import codes.quine.labo.recheck.regexp.PatternExtensions._
 import codes.quine.labo.recheck.unicode.ICharSet
 import codes.quine.labo.recheck.unicode.UString
 import codes.quine.labo.recheck.vm.Program
