+++ modules/recheck/shared/src/main/scala/codes/quine/labo/recheck/vm/ProgramBuilder.scala
@@ -7,6 +7,7 @@ import codes.quine.labo.recheck.common.Context
 import codes.quine.labo.recheck.common.InvalidRegExpException
 import codes.quine.labo.recheck.regexp.Pattern
 import codes.quine.labo.recheck.regexp.Pattern._
+import codes.quine.labo.recheck.regexp.PatternExtensions._
 import codes.quine.labo.recheck.unicode.IChar
 import codes.quine.labo.recheck.unicode.UChar
 import codes.quine.labo.recheck.vm.Inst.AssertKind
